+++ R/calib.functions.R
@@ -36,11 +36,11 @@
 #'
 #'
 #' ## Back-transformation with same calibration
-#' ProxyConversion(proxy.value = as.vector(CalibUk37(temperature = c(21, 22), point.or.sample = "point"))
+#' ProxyConversion(proxy.value = as.vector(ProxyConversion(temperature = c(21, 22), calibration.type = "Uk37", point.or.sample = "point"))
 #'            , point.or.sample = "point", calibration.type = "Uk37")
 #'
 #' ## Back-transformation with random calibration
-#' ProxyConversion(proxy.value = as.vector(CalibUk37(temperature = c(21, 22), point.or.sample = "point"))
+#' ProxyConversion(proxy.value = as.vector(ProxyConversion(temperature = c(21, 22), calibration.type = "Uk37", point.or.sample = "point"))
 #'            , n = 5, point.or.sample = "sample", calibration.type = "Uk37")
 #'
 #' ## Incompatible arguments
@@ -81,7 +81,7 @@ ProxyConversion <- function(temperature = NULL, proxy.value = NULL,
   ## Get calibration parameters
   if (pct != "identity"){
 
-    prs <- calibration.parameters
+    prs <- sedproxy::calibration.parameters
     cfs.vcov <- prs[prs$calibration.type == pct & prs$calibration == calibration, ]
     #if (is.null(calibration)) calibration <- 1
     if (is.null(slp.int.means)){